+++ docs/release_notes.rst
@@ -89,11 +89,15 @@ more details regarding the lower-level changes required to accomplish it.*
 
   - Allow building docker images for a private docker registry
     (`PR#2590 <https://github.com/vantage6/vantage6/pull/2590>`_).
+  - Add community algorithms to the algorithm store by default (`PR#2596
+    <https://github.com/vantage6/vantage6/pull/2596>`_).
 
 - **Change**
 
   - Change the default docker registry to ghcr.io/vantage6, since harbor2.vantage6.ai
     was taken offline (`PR#2587 <https://github.com/vantage6/vantage6/pull/2587>`_).
+  - Allow null values for optional function argument fields in the algorithm store
+    (`PR#2596 <https://github.com/vantage6/vantage6/pull/2596>`_).
 
 4.14.0
 ------
@@ -143,7 +147,6 @@ more details regarding the lower-level changes required to accomplish it.*
     (`Issue#2383 <https://github.com/vantage6/vantage6/issues/2383>`_,
     `PR#2531 <https://github.com/vantage6/vantage6/pull/2531>`_).
 
->>>>>>> parent of e6f474098 (Revert "Merge pull request #2561 from vantage6/release/4.14". This is done so v5.0 can be merged into main)
 
 4.13.7
 ------